Extracted component search extensions - migrated component search extensions to discrete file - added helpers to get and call a method on a collection of components
