+++ Runtime/Scripts/LSL/Utilities/ComponentSearchMethods.cs
@@ -37,50 +37,13 @@ namespace BCIEssentials.Utilities
                 rootObject => rootObject.GetComponentsInChildren<T>(includeInactive)
             ).ToArray();
         }
-    }
-}
 
-namespace BCIEssentials.Extensions
-{
-    using static Utilities.ComponentSearchMethods;
 
-    public static class ComponentSearchExtensions
-    {
-        public static void GetOrAddComponent<T>
+        public static void ForEachComponentInScene<T>
         (
-            this GameObject gameObject,
-            ref T componentReference,
-            bool searchEntireScene = false
-        ) where T : Component
-        {
-            if (
-                componentReference == null &&
-                !gameObject.TryGetComponent(out componentReference) &&
-                (
-                    !searchEntireScene ||
-                    !TryFindComponent(out componentReference)
-                )
-            )
-            {
-                Debug.Log($"No {typeof(T).Name} found, creating one...");
-                componentReference = gameObject.AddComponent<T>();
-            }
-        }
-
-
-        public static bool AnyOther<T>
-        (
-            this T caller,
-            Func<T, bool> predicate,
+            Action<T> action,
             bool includeInactive = false
-        ) where T : Component
-        {
-            FindObjectsInactive findMode = includeInactive
-            ? FindObjectsInactive.Include
-            : FindObjectsInactive.Exclude;
-
-            TryFindComponents(out T[] allComponentsInScene, findMode);
-            return allComponentsInScene.Any(c => c != caller && predicate(c));
-        }
+        )
+        => Array.ForEach(GetComponentsInScene<T>(includeInactive), action);
     }
 }
